[Add]Rspecアルコール画面の編集画面のテスト
